requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
requests==2.31.0
schedule==1.2.1
python-dotenv==1.0.0
pytz==2023.3
APScheduler==3.10.4
flask==2.3.3
gunicorn==21.2.0
python-pushover==0.4 